Cleanup + Font Alterations
Unfortunately, it seems as though the best way to get around this is through not using workshop, seeing as Windows doesn't like non-extracted font files. So, I propose that we link to the github from workshop, that way those who want to use it can get the ZIP file which contains the fonts along with the addon, and at least fix the Papyrus and Comic Sans MS fonts.

The default font is a tricky bastard. I've been messing around with it for hours, but even if I feed it one of the other, already working, fonts, it seems to not work at all. Even though it isn't fixed yet, I recommend we also switch fonts (in this ZIP format version) to a custom one, found here: http://www.fontspace.com/abacadaba/determination-mono-web

On another note, I cleaned up the PSA selection portion with elseif, just to make things more compact, and easier to read.

The only reason I am proposing a change without finishing is because I will be leaving for work soon.